rag-persona-chatbot/pyproject.toml
[tool.poetry]
name = "digitalme-backend"
version = "0.1.0"
description = "RAG-based persona chatbot with intelligent classification"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.35.0"
chromadb = "^0.4.24"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
pyyaml = "^6.0"
numpy = "^1.26.0"
tqdm = "^4.66.0"
structlog = "^24.2.0"
tiktoken = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.0"
isort = "^5.13.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"